+++ test/unit/rmq_2_sysmon/main.py
@@ -27,6 +27,65 @@ import version
 __version__ = version.__version__
 
 
+class ArgParser(object):
+
+    """Class:  ArgParser
+
+    Description:  Class stub holder for gen_class.ArgParser class.
+
+    Methods:
+        __init__
+        arg_dir_chk
+        arg_require
+
+    """
+
+    def __init__(self):
+
+        """Method:  __init__
+
+        Description:  Class initialization.
+
+        Arguments:
+
+        """
+
+        self.cmdline = None
+        self.args_array = dict()
+        self.opt_req = None
+        self.opt_req2 = True
+        self.dir_perms_chk = None
+        self.dir_perms_chk2 = True
+
+    def arg_dir_chk(self, dir_perms_chk):
+
+        """Method:  arg_dir_chk
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_dir_chk.
+
+        Arguments:
+
+        """
+
+        self.dir_perms_chk = dir_perms_chk
+
+        return self.dir_perms_chk2
+
+    def arg_require(self, opt_req):
+
+        """Method:  arg_require
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_require.
+
+        Arguments:
+
+        """
+
+        self.opt_req = opt_req
+
+        return self.opt_req2
+
+
 class UnitTest(unittest.TestCase):
 
     """Class:  UnitTest
@@ -37,10 +96,11 @@ class UnitTest(unittest.TestCase):
         setUp
         test_help_true
         test_help_false
-        test_require_true_chk_true
-        test_require_false_chk_true
-        test_require_true_chk_false
-        test_require_false_chk_false
+        test_require_false
+        test_require_true
+        test_dir_chk_crt_false
+        test_dir_chk_crt_true
+        test_run_program
 
     """
 
@@ -54,10 +114,12 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.args = {"-c": "config_file", "-d": "config_dir", "-M": True}
+        self.args = ArgParser()
+        self.args.args_array = {
+            "-c": "config_file", "-d": "config_dir", "-M": True}
 
     @mock.patch("rmq_2_sysmon.gen_libs.help_func")
-    @mock.patch("rmq_2_sysmon.arg_parser.arg_parse2")
+    @mock.patch("rmq_2_sysmon.gen_class.ArgParser")
     def test_help_true(self, mock_arg, mock_help):
 
         """Function:  test_status_true
@@ -74,7 +136,7 @@ class UnitTest(unittest.TestCase):
         self.assertFalse(rmq_2_sysmon.main())
 
     @mock.patch("rmq_2_sysmon.gen_libs.help_func")
-    @mock.patch("rmq_2_sysmon.arg_parser")
+    @mock.patch("rmq_2_sysmon.gen_class.ArgParser")
     def test_help_false(self, mock_arg, mock_help):
 
         """Function:  test_status_false
@@ -85,91 +147,103 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args
+        self.args.opt_req2 = False
+
+        mock_arg.return_value = self.args
         mock_help.return_value = False
-        mock_arg.arg_require.return_value = True
 
         self.assertFalse(rmq_2_sysmon.main())
 
     @mock.patch("rmq_2_sysmon.gen_libs.help_func")
-    @mock.patch("rmq_2_sysmon.arg_parser")
-    def test_require_true_chk_true(self, mock_arg, mock_help):
+    @mock.patch("rmq_2_sysmon.gen_class.ArgParser")
+    def test_require_false(self, mock_arg, mock_help):
 
-        """Function:  test_require_true_chk_true
+        """Function:  test_require_false
 
-        Description:  Test main function with arg_require returns True and
-            arg_dir_chk_crt returns True.
+        Description:  Test with arg_require returns False.
 
         Arguments:
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args
+        self.args.opt_req2 = False
+
+        mock_arg.return_value = self.args
         mock_help.return_value = False
-        mock_arg.arg_require.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = True
 
         self.assertFalse(rmq_2_sysmon.main())
 
     @mock.patch("rmq_2_sysmon.gen_libs.help_func")
-    @mock.patch("rmq_2_sysmon.arg_parser")
-    def test_require_false_chk_true(self, mock_arg, mock_help):
+    @mock.patch("rmq_2_sysmon.gen_class.ArgParser")
+    def test_require_true(self, mock_arg, mock_help):
 
-        """Function:  test_require_false_chk_true
+        """Function:  test_require_true
 
-        Description:  Test main function with arg_require returns False and
-            arg_dir_chk_crt returns True.
+        Description:  Test with arg_require returns True.
 
         Arguments:
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args
+        self.args.dir_perms_chk2 = False
+
+        mock_arg.return_value = self.args
         mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_dir_chk_crt.return_value = True
 
         self.assertFalse(rmq_2_sysmon.main())
 
     @mock.patch("rmq_2_sysmon.gen_libs.help_func")
-    @mock.patch("rmq_2_sysmon.arg_parser")
-    def test_require_true_chk_false(self, mock_arg, mock_help):
+    @mock.patch("rmq_2_sysmon.gen_class.ArgParser")
+    def test_dir_chk_crt_false(self, mock_arg, mock_help):
 
-        """Function:  test_require_true_chk_false
+        """Function:  test_dir_chk_crt_false
 
-        Description:  Test main function with arg_require returns True and
-            arg_dir_chk_crt returns False.
+        Description:  Test with arg_dir_chk_crt returns False.
 
         Arguments:
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args
+        self.args.dir_perms_chk2 = False
+
+        mock_arg.return_value = self.args
         mock_help.return_value = False
-        mock_arg.arg_require.return_value = True
-        mock_arg.arg_dir_chk_crt.return_value = False
 
         self.assertFalse(rmq_2_sysmon.main())
 
-    @mock.patch("rmq_2_sysmon.run_program")
+    @mock.patch("rmq_2_sysmon.run_program", mock.Mock(return_value=True))
     @mock.patch("rmq_2_sysmon.gen_libs.help_func")
-    @mock.patch("rmq_2_sysmon.arg_parser")
-    def test_require_false_chk_false(self, mock_arg, mock_help, mock_run):
+    @mock.patch("rmq_2_sysmon.gen_class.ArgParser")
+    def test_dir_chk_crt_true(self, mock_arg, mock_help):
 
-        """Function:  test_require_false_chk_false
+        """Function:  test_dir_chk_crt_true
 
-        Description:  Test main function with arg_require returns False and
-            arg_dir_chk_crt returns False.
+        Description:  Test with arg_dir_chk_crt returns True.
 
         Arguments:
 
         """
 
-        mock_arg.arg_parse2.return_value = self.args
+        mock_arg.return_value = self.args
         mock_help.return_value = False
-        mock_arg.arg_require.return_value = False
-        mock_arg.arg_dir_chk_crt.return_value = False
-        mock_run.return_value = True
+
+        self.assertFalse(rmq_2_sysmon.main())
+
+    @mock.patch("rmq_2_sysmon.run_program", mock.Mock(return_value=True))
+    @mock.patch("rmq_2_sysmon.gen_libs.help_func")
+    @mock.patch("rmq_2_sysmon.gen_class.ArgParser")
+    def test_run_program(self, mock_arg, mock_help):
+
+        """Function:  test_run_program
+
+        Description:  Test with run_program.
+
+        Arguments:
+
+        """
+
+        mock_arg.return_value = self.args
+        mock_help.return_value = False
 
         self.assertFalse(rmq_2_sysmon.main())